feat: slicing login page design
