+++ src/modules/LoginModule/index.tsx
@@ -1,107 +1,28 @@
 'use client'
 
-import { useForm } from 'react-hook-form'
-import { zodResolver } from '@hookform/resolvers/zod'
-import * as z from 'zod'
-import { Input } from '@/components/ui/input'
-import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form'
-import { useAuthContext } from '@/contexts/AuthContext'
-import { useState } from 'react'
-import { useRouter } from 'next/navigation'
-import { toast } from 'sonner'
-
-const loginSchema = z.object({
-  email: z.string().email({ message: 'Email tidak valid' }),
-  password: z.string().min(6, { message: 'Password minimal 6 karakter' }),
-})
+import { getImage } from '@/utils/getImage'
+import Image from 'next/image'
+import { JoinUsSection } from '../RegisterModule/sections/JoinUsSection'
+import { RegisterFormHeader } from '../RegisterModule/sections/RegisterFormHeader'
+import { LoginForm } from './module-elements/LoginForm'
 
 export default function LoginModule() {
-  const { login } = useAuthContext()
-  const router = useRouter()
-
-  const [loading, setLoading] = useState<boolean>(false)
-
-  const form = useForm<z.infer<typeof loginSchema>>({
-    resolver: zodResolver(loginSchema),
-    defaultValues: { email: '', password: '' },
-  })
-
-  const onSubmit = async (values: z.infer<typeof loginSchema>) => {
-    console.log('Login values:', values)
-    try {
-      setLoading(true)
-      const response = await login(values)
-      console.log(response)
-      router.push('/')
-      toast.success('Berhasil login!')
-    } catch (err: any) {
-      toast.error((err as Error).message)
-    } finally {
-      setLoading(false)
-    }
-  }
-
   return (
-    <Card className="w-[350px] mx-auto mt-10">
-      <CardHeader>
-        <CardTitle>Masuk</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <Form {...form}>
-          <form
-            name="loginForm"
-            onSubmit={form.handleSubmit(onSubmit)}
-            className="space-y-4"
-          >
-            <FormField
-              control={form.control}
-              name="email"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Email</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="email"
-                      placeholder="Masukkan email"
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="password"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Password</FormLabel>
-                  <FormControl>
-                    <Input
-                      type="password"
-                      placeholder="Masukkan password"
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <Button type="submit" disabled={loading} className="w-full">
-              Masuk
-            </Button>
-          </form>
-        </Form>
-      </CardContent>
-    </Card>
+    <div className="relative">
+      <Image
+        src={getImage('auth_bg.png')}
+        alt="Section Background"
+        fill
+        className="absolute inset-0 object-cover z-0"
+      />
+
+      <section className="flex flex-row relative">
+        <JoinUsSection />
+        <div className="bg-white min-w-[50%] max-w-[50%] flex flex-col items-center justify-center gap-10 p-[2%]">
+          <RegisterFormHeader />
+          <LoginForm />
+        </div>
+      </section>
+    </div>
   )
 }
